# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface User {
  id: string;
  email: string;
  createdAt: string;
}

export interface UserResponse {
  user: User;
}

export interface AuthResponse {
  token: string;
  user: User;
}